Update script to patch openapi definition and push collection contents
